basePrice double -> long

+++ common/src/main/java/com/ssscloud/auction/common/model/Electronic.java
@@ -10,13 +10,13 @@ public class Electronic extends Item {
     private LocalDate purchaseDate;
     private Period warrantyPeriod;
     // constructor ko co 2 thuoc tinh rieng cua Electronic vi chi khi mua hang moi xuat hien trang thai day nen lat nua viet o setter
-    public Electronic (String sellerId, String name, double basePrice, LocalDate manufacturingDate,String creator, String description, boolean isRepaired, Period warrantyPeriod){        
+    public Electronic (String sellerId, String name, long basePrice, LocalDate manufacturingDate,String creator, String description, boolean isRepaired, Period warrantyPeriod){        
         super(sellerId, name, basePrice, manufacturingDate, creator, description);
         this.purchaseDate = null;
         this.isRepaired = isRepaired;
         this.warrantyPeriod = warrantyPeriod;
     }
-    public Electronic (String sellerId, String name, double basePrice, LocalDate manufacturingDate,String creator, String description, boolean isRepaired){        
+    public Electronic (String sellerId, String name, long basePrice, LocalDate manufacturingDate,String creator, String description, boolean isRepaired){        
         super(sellerId, name, basePrice, manufacturingDate, creator, description);
         this.purchaseDate = null;
         this.isRepaired = isRepaired;
